[addons][game] spearate "C" parts to own file
This place the "C" parts about game to a new file on "kodi/c-api/addon-instance/game.h".
This done to allow also "C" language only (as base for other languages) and to confirm safe "C" ABI
between addon and Kodi.
